+++ Sources/StreamChatSwiftUI/ChatChannel/Composer/MessageComposerView.swift
@@ -14,6 +14,7 @@ public struct MessageComposerView<Factory: ViewFactory>: View, KeyboardReadable 
     @State private var popupSize: CGFloat = 350
     @State private var composerHeight: CGFloat = 0
     @State private var keyboardShown = false
+    @State private var editedMessageWillShow = false
     
     private var factory: Factory
     private var channelConfig: ChannelConfig?
@@ -136,13 +137,14 @@ public struct MessageComposerView<Factory: ViewFactory>: View, KeyboardReadable 
         }
         .onReceive(keyboardWillChangePublisher) { visible in
             if visible && !keyboardShown {
-                if viewModel.composerCommand == nil {
+                if viewModel.composerCommand == nil && !editedMessageWillShow {
                     withAnimation(.easeInOut(duration: 0.02)) {
                         viewModel.pickerTypeState = .expanded(.none)
                     }
                 }
             }
             keyboardShown = visible
+            editedMessageWillShow = false
         }
         .onReceive(keyboardHeight) { height in
             if height > 0 && height != popupSize {
@@ -172,6 +174,10 @@ public struct MessageComposerView<Factory: ViewFactory>: View, KeyboardReadable 
         }
         .onChange(of: editedMessage) { _ in
             viewModel.text = editedMessage?.text ?? ""
+            if editedMessage != nil {
+                becomeFirstResponder()
+                editedMessageWillShow = true
+            }
         }
         .accessibilityElement(children: .contain)
     }
@@ -274,7 +280,8 @@ public struct ComposerInputView<Factory: ViewFactory>: View {
                     selectedRangeLocation: $selectedRangeLocation,
                     placeholder: isInCooldown ? L10n.Composer.Placeholder.slowMode : L10n.Composer.Placeholder.message,
                     editable: !isInCooldown,
-                    maxMessageLength: maxMessageLength
+                    maxMessageLength: maxMessageLength,
+                    currentHeight: textFieldHeight
                 )
                 .accessibilityIdentifier("ComposerTextInputView")
                 .accessibilityElement(children: .contain)