+++ StreamChatSwiftUITests/Tests/ChatChannel/MessageComposerView_Tests.swift
@@ -5,6 +5,7 @@
 import SnapshotTesting
 @testable import StreamChat
 @testable import StreamChatSwiftUI
+import SwiftUI
 import XCTest
 
 class MessageComposerView_Tests: StreamChatTestCase {
@@ -114,7 +115,8 @@ class MessageComposerView_Tests: StreamChatTestCase {
             height: .constant(38),
             selectedRangeLocation: .constant(3),
             placeholder: "Send a message",
-            editable: true
+            editable: true,
+            currentHeight: 38
         )
         .frame(width: defaultScreenSize.width, height: 50)
                 
@@ -129,7 +131,8 @@ class MessageComposerView_Tests: StreamChatTestCase {
             height: .constant(38),
             selectedRangeLocation: .constant(3),
             placeholder: "Send a message",
-            editable: true
+            editable: true,
+            currentHeight: 38
         )
         let inputView = InputTextView(
             frame: .init(x: 16, y: 16, width: defaultScreenSize.width - 32, height: 50)
@@ -151,7 +154,8 @@ class MessageComposerView_Tests: StreamChatTestCase {
             height: .constant(38),
             selectedRangeLocation: .constant(3),
             placeholder: "Send a message",
-            editable: true
+            editable: true,
+            currentHeight: 38
         )
         let inputView = InputTextView(
             frame: .init(x: 16, y: 16, width: defaultScreenSize.width - 32, height: 50)
@@ -183,5 +187,29 @@ class MessageComposerView_Tests: StreamChatTestCase {
         
         // Then
         assertSnapshot(matching: view, as: .image)
+    }
+    
+    func test_composerInputView_snapshot() {
+        // Given
+        let inputView = InputTextView()
+        let view = ComposerTextInputView(
+            text: .constant("test test"),
+            height: .constant(100),
+            selectedRangeLocation: .constant(0),
+            placeholder: "Send a message",
+            editable: true,
+            currentHeight: 36
+        )
+        let coordinator = ComposerTextInputView.Coordinator(textInput: view, maxMessageLength: nil)
+        let viewWithSize = view.applyDefaultSize()
+        
+        // When
+        inputView.scrollToBottom()
+        coordinator.updateHeight(inputView, shouldAnimate: true)
+        coordinator.updateHeight(inputView, shouldAnimate: false)
+        
+        // Then
+        assertSnapshot(matching: viewWithSize, as: .image)
+        XCTAssert(coordinator.textInput.height == 100)
     }
 }